requirements.txt
certifi==2019.6.16
chardet==3.0.4
cx-Oracle==7.2.0
gevent==1.4.0
greenlet==0.4.15
grequests==0.3.0
icalendar==4.0.3
idna==2.8
pyproj==2.2.1
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
